# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monitoring Platform API",
    "description": "API for system monitoring, alerts, notifications, and reporting",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/profile/status": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displayUserStatus",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/profile/secureStatus": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displaySecureUserStatus",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/profile/secureDisplay": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displaySecureUserProfile",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/profile/display": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displayUserProfile",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/template": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateTemplate",
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "New Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/secureTemplate": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateSecureTemplate",
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "New Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/secureGenerate": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateSecureNotification",
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "New Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/generate": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateNotification",
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "New Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/message/secureDisplay": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displaySecureMessage",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/message/format": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "formatMessage",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/message/escape": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "escapeMessage",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/message/display": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "displayMessage",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/secureGreeting": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateSecureDashboard",
        "parameters": [
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/greeting": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "operationId": "generateDashboard",
        "parameters": [
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Welcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}